Extract wpAssertion to WordPressShared
