+++ WordPress/Classes/RemotePost+Metadata.swift
@@ -1,4 +1,5 @@
 import Foundation
+import WordPressShared
 
 @objc
 extension RemotePost {
