+++ WordPress/Classes/Utility/TrackableError.swift
@@ -1,56 +1,4 @@
 import Foundation
-import Sentry
-import WordPressShared
-
-// MARK: - Assertions
-
-func wpAssert(_ closure: @autoclosure () -> Bool, _ message: StaticString = "–", userInfo: [String: Any]? = nil, file: StaticString = #file, line: UInt = #line) {
-    guard !closure() else {
-        return
-    }
-
-    let filename = (file.description as NSString).lastPathComponent
-    DDLogError("WPAssertionFailure in \(filename)–\(line): \(message)\n\(userInfo ?? [:])")
-
-    WPAnalytics.track(.assertionFailure, properties: {
-        var properties: [String: Any] = [
-            "assertion": "\(filename)–\(line): \(message)"
-        ]
-        for (key, value) in userInfo ?? [:] {
-            properties[key] = value
-        }
-        return properties
-    }())
-
-    if WPAssertion.shouldSendAssertion(withID: "\(filename)-\(line)") {
-        WPLoggingStack.shared.crashLogging.logError(NSError(domain: "WPAssertionFailure", code: -1, userInfo: [NSDebugDescriptionErrorKey: "\(filename)–\(line): \(message)"]), userInfo: userInfo)
-    }
-
-#if DEBUG
-    assertionFailure(message.description + "\n\(userInfo ?? [:])", file: file, line: line)
-#endif
-}
-
-func wpAssertionFailure(_ message: StaticString, userInfo: [String: Any]? = nil, file: StaticString = #file, line: UInt = #line) {
-    wpAssert(false, message, userInfo: userInfo, file: file, line: line)
-}
-
-private enum WPAssertion {
-    /// The minimum delay between the reports for the same assetion.
-    static let assertionDelay: TimeInterval = 7 * 86400
-
-    static func shouldSendAssertion(withID assertionID: String) -> Bool {
-        let key = "WPAssertionLastReportDateKey-" + assertionID
-        if let lastReportDate = UserDefaults.standard.object(forKey: key) as? Date,
-           Date().timeIntervalSince(lastReportDate) < WPAssertion.assertionDelay {
-            return false
-        }
-        UserDefaults.standard.set(Date(), forKey: key)
-        return true
-    }
-}
-
-// MARK: - Error Tracking
 
 protocol TrackableErrorProtocol {
     func getTrackingUserInfo() -> [String: String]?
