+++ WordPress/Classes/ViewRelated/Blog/My Site/MySiteViewController+FAB.swift
@@ -1,5 +1,6 @@
 import UIKit
 import SwiftUI
+import WordPressShared
 
 extension MySiteViewController {
