+++ WordPress/Classes/ViewRelated/Comments/Controllers/Create/CommentCreateViewModel.swift
@@ -1,5 +1,6 @@
 import Foundation
 import CoreData
+import WordPressShared
 
 final class CommentCreateViewModel {
     var title: String {
