+++ WordPress/Classes/ViewRelated/Comments/Views/Detail/CommentCellViewModel.swift
@@ -1,5 +1,6 @@
 import UIKit
 import Combine
+import WordPressShared
 
 final class CommentCellViewModel: NSObject {
     @objc let comment: Comment
