+++ WordPress/Classes/ViewRelated/Post/Utils/PostNoticeNavigationCoordinator.swift
@@ -1,5 +1,6 @@
 import UIKit
 import SwiftUI
+import WordPressShared
 
 /// This class simply exists to coordinate the display of various sections of
 /// the app in response to actions taken by the user from Post notifications.
