+++ WordPress/Classes/ViewRelated/System/Sidebar/SidebarViewModel.swift
@@ -2,6 +2,7 @@ import Combine
 import Foundation
 import WordPressData
 import WordPressKit
+import WordPressShared
 
 enum SidebarSelection: Hashable {
     case welcome
